+++ app/src/main/java/com/sopt/now/compose/data/Friend.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.data
 
 import androidx.annotation.DrawableRes
