+++ app/src/main/java/com/sopt/now/compose/data/User.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.data
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
