+++ app/src/main/java/com/sopt/now/compose/ui/main/MainActivity.kt
@@ -1,7 +1,6 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.ui.main
 
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.Arrangement
@@ -26,6 +25,10 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.sopt.now.compose.BottomNavigationItem
+import com.sopt.now.compose.data.User
+import com.sopt.now.compose.ui.home.HomeScreen
+import com.sopt.now.compose.ui.mypage.MyPageScreen
 import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 
 class MainActivity : ComponentActivity() {
